requirements.txt
gitpython
githubcrawler
stdlib-list
pip
PyGithub
python-magic
asynctcp
redis
rq
rsyslog
